fix wrong terminal size on windows
Looks like this commit broke getting the terminal size on windows:

342c47eb1ba371d37b9cbd1a1db85f7bbe3b6ad

They modified the size of the `wAttributes` field on `CONSOLE_SCREEN_BUFFER_INFO`.
This appears to have been a mistake as it looks like they probably just
did a search/replace of all the `u16` types with `u32`, however, we
can't change the size of this field as it's defined by the OS and
in doing so our `srWindow` offset is now incorrect.